api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic-settings>=2.1.0
httpx>=0.26.0
